Attempt to generalise base component
