+++ ServiceWebsite/ServiceWebsite/ClientApp/src/app/modules/individual-journey/components/suitability-choice-page-base.component.spec.ts
@@ -1,38 +1,10 @@
-import { ComponentFixture } from '@angular/core/testing';
-import { ConfigureTestBedFor } from '../pages/individual-base-component/component-test-bed.spec';
+import { ConfigureTestBedFor } from '../../individual-journey/pages/individual-base-component/component-test-bed.spec';
+import { Type } from '@angular/core';
 import { SuitabilityChoicePageBaseComponent } from './suitability-choice-page-base.component';
-import { Type, DebugElement } from '@angular/core';
-import { By } from '@angular/platform-browser';
-
-class SuitabilityChoicePageBaseFixture<T extends SuitabilityChoicePageBaseComponent> {
-  readonly fixture: ComponentFixture<T>;
-  readonly component: T;
-
-  constructor(fixture: ComponentFixture<T>) {
-    this.fixture = fixture;
-    this.component = fixture.componentInstance;
-  }
-
-  detectChanges(): void {
-    this.fixture.detectChanges();
-  }
-
-  radioBoxIsClicked(id: string) {
-    const radioButton = this.debugElementByCss(id);
-    radioButton.nativeElement.click();
-    this.fixture.detectChanges();
-  }
-
-  submitIsClicked() {
-    const continueButton = this.debugElementByCss('.govuk-button');
-    continueButton.nativeElement.click();
-    this.fixture.detectChanges();
-  }
-
-  debugElementByCss(css: string): DebugElement {
-    return this.fixture.debugElement.query(By.css(css));
-  }
-}
+import {
+  CannotProceeedUntilChoiceIsSelected as cannotProceedUntilChoiceIsSelectedBase,
+  SuitabilityChoicePageBaseFixture
+} from '../../base-journey/components/suitability-choice-page-base.component.spec';
 
 const configureTestBedFor = <T extends SuitabilityChoicePageBaseComponent>
   (component: Type<T>, customiseConfiguration?: Function): SuitabilityChoicePageBaseFixture<T> => {
@@ -50,28 +22,7 @@ const cannotProceedUntilChoiceIsSelected =
   <T extends SuitabilityChoicePageBaseComponent>(component: Type<T>, customiseConfiguration?: Function):
     SuitabilityChoicePageBaseFixture<T> => {
     const fixture = configureTestBedFor(component, customiseConfiguration);
-
-    // when
-    fixture.submitIsClicked();
-
-    // then
-    expect(fixture.component.isFormInvalid).toBeTruthy();
-
-    // expect form to be erronous
-    const formContainer = fixture.debugElementByCss('#form-container');
-    expect(formContainer.classes['govuk-form-group--error']).toBeTruthy();
-
-    // and error message to be displayed
-    const errorMessage = fixture.debugElementByCss('#error-message');
-    expect(errorMessage.nativeElement).toBeTruthy();
-
-    // when selecting a radio button
-    fixture.radioBoxIsClicked('#choice-yes');
-
-    // then
-    fixture.submitIsClicked();
-
-    return fixture;
+    return cannotProceedUntilChoiceIsSelectedBase(fixture, component);
   };
 
 export {
